requirements.txt
flask>=2.0
pycryptodome>=3.20.0
streamlit>=1.0
requests>=2.28
